requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
ffmpeg-python>=0.2.0
tqdm>=4.65.0
openai-whisper>=20231117
jiwer>=3.0.0
pandas>=2.0.0
librosa>=0.10.0
soundfile>=0.12.1
matplotlib>=3.7.0
datasets>=2.18,<4.0
transformers>=5.0.0
huggingface-hub>=1.3.0,<1.4.0
accelerate>=1.12.0
peft>=0.18.0
PyYAML>=6.0
evaluate>=0.4.0
Pillow>=10.0.0
torch>=2.0.0
